+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Changelog](https://keepachangelog.com/en/1.1.0).
   * NFT token name for Checkpoint and Committee Oracle set to empty string.
     Init token name for Committee Oracle set to `Committee oracle InitToken`.
     Previuous values were invalid due to using wrong parsing function.
+  * Some `String` types changed to strict `ByteString` where appropriate.
 
 ## Removed